abstracted api call in use effect to server component
